Move ReadOnlyDB trait into backend
